% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ecoregions_species.R
\name{dl_ecoregions_species_data}
\alias{dl_ecoregions_species_data}
\title{Download ecoregions species dataset}
\usage{
dl_ecoregions_species_data(overwrite = FALSE)
}
\arguments{
\item{overwrite}{a \code{logical}. If \code{TRUE}, the file will be downloaded again
and the previous version will be replaced.}
}
\value{
No return value.
}
\description{
This function downloads the ecoregions species dataset (csv file) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}. The file
won't be downloaded if already exists locally (except if \code{overwrite = TRUE}).
The file \code{wildfinder-ecoregions_species.csv} will be stored in
\verb{data/wildfinder/}. This folder will be created if required.
}
