% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_pantheria_data.R
\name{dl_pantheria_data}
\alias{dl_pantheria_data}
\title{Download  pantheria data}
\usage{
dl_pantheria_data()
}
\value{
This function return the downloaded file path PanTHERIA_1-0_WR05_Aug2008.txt
}
\description{
\code{dl_pantheria_data}downloads the file
}
